backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
passlib[bcrypt]
python-multipart
python-jose[cryptography]
python-dotenv
requests
